refactor: move all unused services and routes to legacy and update tests
